+++ Modules/Intent.Modules.IaC.Terraform/Templates/Applications/AzureEventGridResourcesTf/AzureEventGridResourcesTfTemplatePartial.cs
@@ -52,18 +52,62 @@ namespace Intent.Modules.IaC.Terraform.Templates.Applications.AzureEventGridReso
         {
             var builder = new TerraformFileBuilder();
 
-            var items = new HashSet<string>();
-            foreach (var message in _azureEventGridMessages)
+            // Group messages by domain
+            var messagesByDomain = _azureEventGridMessages
+                .GroupBy(m => m.DomainName)
+                .ToList();
+
+            foreach (var domainGroup in messagesByDomain)
             {
-                if (items.Add(message.TopicName))
+                if (domainGroup.Key != null)
                 {
-                    builder.AddResource(Terraform.azurerm_eventgrid_topic.type, Terraform.azurerm_eventgrid_topic.topic(message).refname, resource =>
+                    // Event Domain pattern - generate domain + domain topics
+                    var domainName = domainGroup.Key;
+                    var domainData = Terraform.azurerm_eventgrid_domain.domain(domainName);
+                    
+                    // Generate the domain resource
+                    builder.AddResource(Terraform.azurerm_eventgrid_domain.type, domainData.refname, resource =>
                     {
-                        resource.AddSetting("name", message.TopicName);
+                        resource.AddSetting("name", $"{domainName}");
                         resource.AddRawSetting("location", Terraform.azurerm_resource_group.main_rg.location);
                         resource.AddRawSetting("resource_group_name", Terraform.azurerm_resource_group.main_rg.name);
                         resource.AddSetting("input_schema", "CloudEventSchemaV1_0");
                     });
+
+                    // Generate domain topic resources for each unique topic in the domain
+                    var uniqueTopics = new HashSet<string>();
+                    foreach (var message in domainGroup)
+                    {
+                        if (uniqueTopics.Add(message.TopicName))
+                        {
+                            var domainTopicData = Terraform.azurerm_eventgrid_domain_topic.domainTopic(message);
+                            builder.AddResource(Terraform.azurerm_eventgrid_domain_topic.type, domainTopicData.refname, resource =>
+                            {
+                                resource.AddSetting("name", message.TopicName);
+                                resource.AddRawSetting("resource_group_name", Terraform.azurerm_resource_group.main_rg.name);
+                                resource.AddRawSetting("domain_name", $"{Terraform.azurerm_eventgrid_domain.type}.{domainData.refname}.name");
+                            });
+                        }
+                    }
+                }
+                else
+                {
+                    // Custom Topic pattern - generate individual topics (backwards compatibility)
+                    var uniqueTopics = new HashSet<string>();
+                    foreach (var message in domainGroup)
+                    {
+                        if (uniqueTopics.Add(message.TopicName))
+                        {
+                            var topicData = Terraform.azurerm_eventgrid_topic.topic(message);
+                            builder.AddResource(Terraform.azurerm_eventgrid_topic.type, topicData.refname, resource =>
+                            {
+                                resource.AddSetting("name", message.TopicName);
+                                resource.AddRawSetting("location", Terraform.azurerm_resource_group.main_rg.location);
+                                resource.AddRawSetting("resource_group_name", Terraform.azurerm_resource_group.main_rg.name);
+                                resource.AddSetting("input_schema", "CloudEventSchemaV1_0");
+                            });
+                        }
+                    }
                 }
             }