+++ Tests/AzureFunctions.AzureEventGrid.GroupA/AzureFunctions.AzureEventGrid.Infrastructure/Configuration/AzureEventGridConfiguration.cs
@@ -1,4 +1,5 @@
 using AzureFunctions.AzureEventGrid.Application.Common.Eventing;
+using AzureFunctions.AzureEventGrid.EventDomain;
 using AzureFunctions.AzureEventGrid.GroupA.Eventing.Messages;
 using AzureFunctions.AzureEventGrid.GroupB.Eventing.Messages;
 using AzureFunctions.AzureEventGrid.Infrastructure.Eventing;
@@ -28,7 +29,11 @@ namespace AzureFunctions.AzureEventGrid.Infrastructure.Configuration
             services.AddScoped<IAzureEventGridConsumerBehavior, InboundCloudEventBehavior>();
             services.Configure<AzureEventGridPublisherOptions>(options =>
             {
-                options.Add<ClientCreatedEvent>(configuration["EventGrid:Topics:ClientCreatedEvent:Key"]!, configuration["EventGrid:Topics:ClientCreatedEvent:Endpoint"]!, configuration["EventGrid:Topics:ClientCreatedEvent:Source"]!);
+                options.AddTopic<ClientCreatedEvent>(configuration["EventGrid:Topics:ClientCreatedEvent:Key"]!, configuration["EventGrid:Topics:ClientCreatedEvent:Endpoint"]!, configuration["EventGrid:Topics:ClientCreatedEvent:Source"]!);
+                options.AddDomain(configuration["EventGrid:Domains:MainDomain:Key"]!, configuration["EventGrid:Domains:MainDomain:Endpoint"]!, domain =>
+                {
+                    domain.Add<OrderCreatedEvent>(configuration["EventGrid:Topics:OrderCreatedEvent:Source"]!);
+                });
             });
             services.Configure<AzureEventGridSubscriptionOptions>(options =>
             {