+++ Tests/AzureFunctions.AzureEventGrid.GroupA/AzureFunctions.AzureEventGrid.Infrastructure/Configuration/AzureEventGridPublisherOptions.cs
@@ -13,12 +13,47 @@ namespace AzureFunctions.AzureEventGrid.Infrastructure.Configuration
 
         public IReadOnlyList<PublisherEntry> Entries => _entries;
 
-        public void Add<TMessage>(string credentialKey, string endpoint, string source)
+        public void AddTopic<TMessage>(string credentialKey, string endpoint, string source)
         {
             ArgumentNullException.ThrowIfNull(credentialKey);
             ArgumentNullException.ThrowIfNull(endpoint);
             ArgumentNullException.ThrowIfNull(source);
             _entries.Add(new PublisherEntry(typeof(TMessage), credentialKey, endpoint, source));
+        }
+
+        public void AddDomain(string credentialKey, string endpoint, Action<DomainOptions>? domainAction)
+        {
+            ArgumentNullException.ThrowIfNull(credentialKey);
+            ArgumentNullException.ThrowIfNull(endpoint);
+
+            var domainOptions = new DomainOptions(credentialKey, endpoint);
+            domainAction?.Invoke(domainOptions);
+
+            _entries.AddRange(domainOptions.ToEntries());
+        }
+    }
+
+    public class DomainOptions
+    {
+        private readonly string _credentialKey;
+        private readonly string _endpoint;
+        private readonly List<PublisherEntry> _entries = [];
+
+        public DomainOptions(string credentialKey, string endpoint)
+        {
+            _credentialKey = credentialKey;
+            _endpoint = endpoint;
+        }
+
+        public void Add<TMessage>(string source)
+        {
+            ArgumentNullException.ThrowIfNull(source);
+            _entries.Add(new PublisherEntry(typeof(TMessage), _credentialKey, _endpoint, source));
+        }
+
+        public IEnumerable<PublisherEntry> ToEntries()
+        {
+            return _entries;
         }
     }