+++ Tests/AzureFunctions.AzureEventGrid.GroupB/AzureFunctions.AzureEventGrid.Application/IntegrationEvents/OrderCreatedEvent.cs
@@ -1,0 +1,17 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.Contracts.IntegrationEventMessage", Version = "1.0")]
+
+namespace AzureFunctions.AzureEventGrid.EventDomain
+{
+    public record OrderCreatedEvent
+    {
+        public OrderCreatedEvent()
+        {
+            RefNo = null!;
+        }
+
+        public string RefNo { get; init; }
+    }
+}